[spirv] Switch to use common target description (#17623)
This commit switches SPIR-V side to use the common
`#iree_gpu.target` to describe the GPU characteristics.
With it we can now remove the ad-hoc Vulkan attributes and
dialects and unify how GPU are described across various GPU
compiler backends in IREE.

#### SPIR-V

SPIR-V has some additional requirements that we need to
account for:

We have many vendors and APIs to handle there so this commit
adds various AMD/ARM/NVIDIA/Qualcomm targets for
development purposes so that we can specify them with
a shorthand.

In order to be extensible, leverage the `feature` field in
`#iree_gpu.target` to specify additional capabilities with
`cap:` prefix and extensions with `ext:` prefix. We also
use the `feature` field to specify what SPIR-V version to
target with the `spirv:v1.x` format.

Right now the `SPIRVConvertGPUTarget` pass is
invoked immediately before configuration. This is to
stage the changes. As a next step we need to move
it immediately before `ConvertToSPIRV` pass.

#### Vulkan

`--iree-vulkan-target-env` is dropped given now we
removed the whole Vulkan dialect and cannot control with
a `#vk.target_env` attribute anymore.

The default `--iree-vulkan-target-triple` now becomes
`vp_android_baseline_2022`, which is a a good lowest
common denominator to guarantee the generated SPIR-V
is widely accepted. We are not considering SwiftShader
now anymore like previously due to testing purposes.

The `--iree-vulkan-target-triple` should be renamed
given it's not a triple anymore--that will happen later
together with other GPU backends (i.e., cuda/hip) to
be consistent.

In order to support cooperative matrix conversion, we
added `WMMA_F16_16x16x16_F16`. For NVIDIA GPUs
we are abusing it right now without considering the
concrete explicit layout--that is fine given in Vulkan
they are opaque anyway. But this need to be fixed if
we are targeting WMMA in CUDA.

#### Metal / WebGPU

We now contruct a `#iree_gpu.target` to specify
the target to drive SPIR-V CodeGen.

Progress towards https://github.com/iree-org/iree/issues/16341

ci-extra: test_nvidia_gpu,test_nvidia_a100,test_amd_mi250,
build_test_all_macos_arm64,build_and_test_android