Change `EncodingRole` to `IntegerAttr` (#17708)
In the future it is possible to encode types other than a matmul, so
changing the `EncodingRole` to be the index of the operand. Instead of
using integers, use enum to retain the role's symbolic meaning in the
code.


Note that: It is one's responsibility to make sure the operand index is
correct. The index follows the convention that in MLIR output operands
are after input operands and are at the end of the operand list.

Signed-off-by: Alan Li <me@alanli.org>