[Im2col] Add decomposition pass for iree_linalg_ext.im2col (#17728)
This PR adds a pass for decomposing the `iree_linalg_ext.im2col` op into
serial loops of `extract_slice->copy->insert_slice`. The pass tries to
keep the innermost dimension of the iteration space un-tiled when the
inner slice is contiguous, but otherwise falls back to serial loops of
scalar slices. The outer dimensions (`B` and `M`) of the iteration
space, however, are always tiled to 1.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>