[NFC] Rename `FusionOfTensorOps` to `FuseMultiUseElementwiseProducer`. (#17828)
This was a refactoring that was left out of previous refactoring of
elementwise operation fusion. The `FusionOfTensorOps` pass was only
handling fusing elementwise operations where producer had multiple uses.
Rename the pass to make this clearer. Also split/rename the tests as
well to make the mapping of test to passes/pipeline clearer.

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>