Drop unmaintained transform dialect tests (#17858)
The tests in `tests/transform_dialect` have been unmaintained (and
mostly disabled) for a while. Most of the tests were misplaced anyway,
mixing lit tests in with e2e tests, and much of the relevant
functionality is now tested upstream or has evolved into different lit
tests in Codegen.

This drops the entire `cuda` directory, and drops all `cpu` tests except
one which is testing a transform dialect library call, which is still a
relevant e2e test.