Disable arm64 runtime build while runner availability is low. (#17927)
We currently only have a single `t2a-standard-8` runner for arm64
presubmit jobs. The jobs we have are:

job name | when does it run?
-- | --
`build_test_runtime :: arm64` | all PRs
`build_test_all_arm64` | opt-in, always enabled on LLVM integrate PRs

While the runtime job is always fast (~4 minutes), the 'all' (compiler)
job takes 15 minutes with a warm cache and 1h30m with a cold cache.
Having only a single runner results in presubmit jobs getting starved
for up to several times our presubmit target (20-30 minutes).

For now this just disables the runtime job entirely, trusting that
changes affecting it will be infrequent. I also explored options for
making it opt-in, but support for conditional jobs in matrices within
GitHub Actions is a complete mess. We could split this back out into a
non-matrix job is we want to go down that path.