[Codegen][GPU] Use bufferization.alloc_tensor for gpu.shuffle_tensor destination (#17940)
The destination for the gpu.shuffle_tensor op will always end up needing
shared memory allocations. When the destination is left as a
tensor.empty op, it can potentially be CSEd with other
gpu.shuffle_tensor destinations. This PR creates a
bufferization.alloc_tensor when generating gpu.shuffle_tensor ops
instead, which will not be CSEd.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>