[Codegen] Ensure hoisted extraction replaced by induction var. (#17975)
This commit teaches the compiler to replace hoisted extraction of IV by
the newly generated IV with the correct shapes. Previously we would
hoist the extraction and replace the IV uses by the hoisted extraction,
however this may not always be correct since the IV's value may be
updated in the loop.

The main motivation of this PR is to fix numerical issue caused by such
case that exists in the attention-cpp pipeline. Although this happens at
the vector level as opposed to the test cases we have for at tensor
level, we can re-use said test. Specific example for this case will be
left in the comment section of this PR.

---------

Signed-off-by: Stanley Winata <stanley.winata@amd.com>