[Encoding] Add an optional bcast_map attribute to EncodingAttr. (#18032)
This adds a new optional field to encodings called `bcast_map`. When we
set encodings, we want to set encodings on the inputs to broadcasting
operations, since it is less data to pack. This `bcast_map` is a step
towards being able to do this, since it is needed in order to know which
dimensions of the tensor correspond to which dimensions of the packed
layout.

The new field is an affine map that encodes which dimensions of the
encoded tensor map to which dimensions in the corresponding operand of
the data tiled op. For example, if the LHS of a matmul is broadcasted
along the batch dimension, and we set encoding on the input to the
broadcast:
```mlir
#map = affine_map<(d0, d1, d2) -> (d1, d2)>
#map1 = affine_map<(d0, d1, d2) -> (d0, d1, d2)>
%se = iree_encoding.set_encoding %lhs : tensor<4x16xf32> -> tensor<4x16xf32, ... bcast_map = #map
%bcast = linalg.broadcast ins(%lhs ... outs(%e : tensor<2x4x16xf32, ... bcast_map = #map1 ... dimensions = [0]
```
The result of the broadcast, which will be consumed by the matmul, has
an identity broadcast map, and the input to the broadcast has a
broadcasted affine map. The `#map` says that the dimensions of `%se`
correspond to `d1` and `d2` in the LHS of the matmul that consumes
`%bcast`.

In cases where we transpose narrow N matmuls, the `bcast_map` remains
the same. Handling this properly is left as a TODO, to be fixed when
more pieces land, and we can more properly test transposed narrow N
matmuls. This is okay for now, since the `bcast_map` is not actually
used anywhere yet.

Signed-off-by: hanhanW <hanhan0912@gmail.com>
Co-authored-by: hanhanW <hanhan0912@gmail.com>