[VectorDistribution] Replace layout_resolution with to_layout (#18027)
This patch replaces the layout_resolution operator with a new
"to_layout" operation, representing a layout cast on the result. This
allows the operation to be used as an anchor and a conversion operation.

This operation will be used in later patches to set layout anchors in IR
and preserve them.