Removing extraneous iree/hal/drivers/utils. (#18050)
We want to keep the total number of targets reasonable and having a new
target for a single helper is not worth the cost. We also already have a
utils directory meant for HAL implementations to use and it should have
been in there. Since this method is simple it's been added to the core
HAL header (usually not preferred, but in the case of semaphore_list its
a basic utility and not a vtabled type that will have different behavior
between implementations).