[Flow] Fix error in CollapseDimensionsPass (#18128)
There was a crash described in
https://github.com/iree-org/iree/issues/18126. The problem was that I
was getting a reference to `CollapseInfo` and then using the bracket
operator to access the map (MapVector) again, which can modify the map
and invalidate the reference. This was occurring because opMap was not
populated with all the generic ops in the dispatch region (because of
isEligibleForCollapse). The logic to find a producer/consumer looked for
a generic with the same parent, which can include ops not in the map,
causing elements to be added to the map when there was a reference to an
object in the map.

In addition, I changed MapVector to DenseMap because there is no need to
iterate over the map.

---------

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>