[Codegen][GPU] Add pass to expand multi_mma op shapes to intrinsic layout (#18139)
This PR adds a new pass to explicitly materialize the dimensions of
intrinsic layouts for `iree_gpu.multi_mma` ops. This means adding an
expand_shape on each of the inputs to go from the `OpaqueMmaLayout`
shape to the `ConcreteMmaLayout` shape. This makes it easy to extract
the correct data from the tensors when it is time to distribute the
multi_mma op to lanes, since the shape will match the number of offsets
and sizes needed for the slice.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>