Adding split iree_hal_buffer_prepare_map_range/commit_map_range. (#18159)
This allows for the validation/offsetting/etc that acts only on metadata
to occur separate from the actual commit of the mapping where the host
storage is accessed. Code that wants to validate early (such as during
command buffer recording) can now do so on uncommitted buffers and have
a fast-path for when the underlying storage is available.

This will be used for native implementations of reusable command buffers
by allowing them to validate direct and indirect bindings during
recording even if the bindings are of uncommitted buffers (#18158) that
won't be available until execution-time.