[GPU] Add check for contractionOpInterface in setMatmulLoweringConfig (#18178)
The `setMatmulLoweringConfig` function only checks that
`linalg::inferContractionDims` returns contraction dimensions, but not
that the operation is a contraction. The inferContractionDims function
does not check that maps are projected permutations, so it can infer
contraction dims on non-pure contraction ops, like convolutions. This
causes a segmentation fault because the assumption of
`setMatmulLoweringConfig` is that there the op is a pure contraction.
This fixes the bug by checking for `linalg::isaContractionOpInterface`
as well.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>