Bubble expand shapes through `AttentionOp`s (#18074)
The idea was to incrementally land these changes (see
https://github.com/iree-org/iree/pull/18030 and
https://github.com/iree-org/iree/pull/18017), but it appears that they
are dependent on each other to not cause regressions in sdxl fp16. I
will give a summary of what the commits here do.

## 1. Make all input maps identity (originally on PR
https://github.com/iree-org/iree/pull/18017)
First two commits
> Change ElementwiseOpInterchangePattern from making output maps
identity to making the input maps identity.
https://github.com/iree-org/iree/issues/18006

## 2. Add `LinalgFusionOpInterface` to attention op (originally on PR
https://github.com/iree-org/iree/pull/18030)
Next two commits after merge
>- Added interface to attention op so that producers/consumers can get
fused into the same dispatch.
>- Cleaned up interface method naming by changing `getIndexingMaps` to
`getIndexingMapsArray` to better match linalg. `getIndexingMaps` also
already conflicted with a method that attention already has.
>- Added indexing check to attention verifier and corresponding test.

## 3. Most recent commit on this PR
This was going to be the only commit of this PR and is the one that
needs the most review. It creates
LinalgExt::populateFoldReshapeOpsByExpansionPattern that emulates
linalg::populateFoldReshapeOpsByExpansionPattern but for AttentionOp. A
bit more info here https://github.com/iree-org/iree/issues/17673.

---------

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>
Signed-off-by: saienduri <saimanas.enduri@amd.com>
Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>
Co-authored-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>
Co-authored-by: saienduri <saimanas.enduri@amd.com>