[Stream] fix ordering issue in EmplaceAllocations (#18321)
Fix a dominance issue when the updates are from last to first results
for a dispatch with multiple results. This is done by adding two phases
to the pass. First phase collects the update ops that can be potentially
placed in dispatch results. Then we sort and apply them in block order.
Fixes : https://github.com/iree-org/iree/issues/18308