[LLVMGPU] Remove old LLVMGPULayoutAnalysisAndDistribution (#18115)
This patch removes the old layout analysis and distribution. This
backend was only being used by a transform script to compile attention,
but isn't really used anywhere else and is not maintained.

This backend can be completely replaced by the new vector distribution
backend. One downside of removing this pass is we cannot compile
attention for mma.sync anymore, but we can eventually add support for
that in vector distribution (or TileAndFuse) in a much better, non
transform dialect script way.