[Codegen] Resolve constantOp with multiple layouts users. (#18354)
Main motivation is to handle distribution of constantOp who has users
with different layouts.

Original use case is to ensure we can distribute attention when the tile
size for M,K1,N is the same. Which means the init of 1st contract, and
IV's init uses the same constantOp.

Since constantOp can only hold a single layout, but multiple to_layout
ops with different layouts, for each user, there will be non resolved
to_layout op(s). only one of the to_layout op can be resolved properly,
the rest would be a "non trivial" resolution since layouts are
different.

To solve this issue, we introduce a mechanism that detect these cases
and make a copy of the arith.constant that get used by other users, when
we are trying to resolve for the current constantOp.

Signed-off-by: Stanley Winata <stanley.winata@amd.com>