[Codegen][GPU] Support all kinds in combiningKindToAllReduce() (#18443)
Changes this function so it supports all members of the CombiningKind
enum; the min/max discrepancies that originally motivated their
exclusion no longer exist. Since it covers all enum members, it is
changed to no longer return an optional. Note that this changes the
generated MLIR (gpu.subgroup_reduce is generated), but that should
ultimately lower to the same pattern of shuffles as before.

Function is made `static` as it is not used anywhere else and has no
prototype in a header.

Signed-off-by: Andrea Faulds <andrea.faulds@amd.com>