[VectorDistribution] Add support for multi-subgroup attention (#18188)
This patch adds support for distributing attention to multiple
subgroups.

Currently, we distinguish the two matmuls in attention by setting a
discardable attribute on the matmuls (set during decomposition) used as
a hint to layout anchoring, on what to do when it encounters these
matmuls. (Note that even if these hints were dropped, it would only lead
to a drop in performance, because the layout anchoring doesn't know its
attention anymore). The correct way to handle these matmuls would be to
start putting mma_schedule as an operation specific lowering config and
teach decomposition to propagate this lowering to the two matmuls after
decomposition. This is blocked by work on TileAndDistributeToWorkgroups
supporting consumer fusion, and needs some heavy lifting.