[GPU][DT] Add support for GPU data-tiling E2E tests.  (#18591)
The revision introduces GPUMaterializeHostEncodingPass for early
materialization. It is mainly for testing purpose while the codegen for
encodings is still in progress. A experimental flag (i.e.,
`iree-global-opt-experimental-rocm-data-tiling`) is introduced to
provide a path for GPU data-tiling e2e tests.

In the tests, a `optimization_barrier` op is introduced in between of
`set_encoding` op and `unset_encoding` op. The compiler could be clever
to cancel data-layout transformation, so a barrier is inserted as a
hint. To support the e2e path, the revision adds two additional changes:

1. Implement `MaterializeOptimizationBarrierOp` pattern, which replaces
the barrier op with the same op with materialized shape.
2. Implement the fallback for unset_encoding materialization pattern.
Currently only f32.f32.f32 and i8.i8.i32 mfma ops are supported. The
wmma ops are not supported, so the codegen should turn encodings to nop
on the targets that do not yet support the intrinsics. E.g., gfx1100.

Note: this only tests set_encoding and unset_encoding. The gemm codegen
is still in progress.

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>